+++ graph.go
@@ -34,7 +34,7 @@ type graphInput struct {
 }
 
 type graphTask struct {
-	display TaskDisplay
+	display            TaskDisplay
 	id                 string
 	name               string
 	module             string